tambahan

+++ app/Http/Controllers/Api/JenisTransaksiController.php
@@ -5,6 +5,8 @@ namespace App\Http\Controllers\Api;
 use App\Http\Controllers\Controller;
 use App\Http\Resources\JenisTransaksiResource;
 use App\Models\JenisTransaksi;
+use App\Models\PenerimaanLangsung;
+use App\Models\Pengajuan;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Validator;
 
@@ -67,6 +69,14 @@ class JenisTransaksiController extends Controller
     {
         $jenisTransaksi = JenisTransaksi::where('trx_id', $id)->first();
         if ($jenisTransaksi) {
+            $isUsedPenerimaan = PenerimaanLangsung::where('trans_jns', $id)->exists();
+            $isUsedPengajuan = Pengajuan::where('trx_id', $id)->exists();
+            if ($isUsedPenerimaan) {
+                return new JenisTransaksiResource(false, 'Data Jenis Transaksi Sudah digunakan pada Penerimaan Langsung!', null);
+            }
+            if ($isUsedPengajuan) {
+                return new JenisTransaksiResource(false, 'Data Jenis Transaksi Sudah digunakan pada Pengajuan Pengeluaran!', null);
+            }
             $jenisTransaksi->delete();
             return new JenisTransaksiResource(true, 'Data Jenis Transaksi Berhasil Dihapus!', null);
         } else {
